+++ repos/exposed/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/exposed/ExposedCRUDRepo.kt
@@ -3,9 +3,7 @@ package dev.inmo.micro_utils.repos.exposed
 import dev.inmo.micro_utils.repos.Repo
 import org.jetbrains.exposed.sql.*
 
-interface ExposedCRUDRepo<ObjectType, IdType> : Repo {
-    val database: Database
-
+interface ExposedCRUDRepo<ObjectType, IdType> : ExposedRepo {
     val ResultRow.asObject: ObjectType
     val selectById: SqlExpressionBuilder.(IdType) -> Op<Boolean>
 }