+++ repos/exposed/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/exposed/ExposedRepo.kt
@@ -1,0 +1,8 @@
+package dev.inmo.micro_utils.repos.exposed
+
+import dev.inmo.micro_utils.repos.Repo
+import org.jetbrains.exposed.sql.Database
+
+interface ExposedRepo : Repo {
+    val database: Database
+}