+++ repos/exposed/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/exposed/ExposedTableInitialization.kt
@@ -6,3 +6,5 @@ import org.jetbrains.exposed.sql.transactions.transaction
 fun Table.initTable(database: Database) {
     transaction(database) { SchemaUtils.createMissingTablesAndColumns(this@initTable) }
 }
+
+fun <T> T.initTable() where T: ExposedRepo, T: Table = initTable(database)