+++ CHANGELOG.md
@@ -1,10 +1,36 @@
 # Changelog
 
-## 0.1.0
+## 0.2.0
 
-Inited :)
+* `Repos`
+    * `Exposed`
+        * Now this project depend on `micro_utils.coroutines`
+        * Typealias `ColumnAllocator` has been replaced to root exposed package
+        * Interface `ExposedRepo` has been added
+            * `ExposedCRUDRepo` now extends `ExposedRepo` instead of simple `Repo`
+        * New extension `initTable` for classes which are `Table` and `ExposedRepo` at the same time
+        * `KeyValue`:
+            * Class `AbstractExposedReadKeyValueRepo`
+                * Renamed to `ExposedReadKeyValueRepo`
+                * Changed incoming types to `ColumnAllocator`
+                * `open` instead of `abstract`
+                * Implements `ExposedRepo`
+            * Class `AbstractExposedKeyValueRepo`
+                * Renamed to `ExposedKeyValueRepo`
+                * Changed incoming types to `ColumnAllocator`
+                * `open` instead of `abstract`
+        * `OneToMany`:
+            * Class `AbstractExposedReadOneToManyKeyValueRepo`
+                * Renamed to `ExposedReadOneToManyKeyValueRepo`
+                * Changed incoming arguments order
+                * Implements `ExposedRepo`
+            * Class `AbstractExposedOneToManyKeyValueRepo`
+                * Renamed to `ExposedKeyValueRepo`
+                * Changed incoming arguments order
+                * `open` instead of `abstract`
+                * Release for every `Flow` in parent interfaces
 
-### 0.1.1
+## 0.1.1
 
 * `Versions`:
     * `kotlinx.serialization`: `1.0.0-RC2` -> `1.0.0`
@@ -39,3 +65,8 @@ Inited :)
         * `MapCRUDRepo` class as implementation of `StandardCRUDRepo` on top of `MutableMap` has been added
         * `MapKeyValueRepo` class as implementation of `StandardKeyValueRepo` on top of `MutableMap` has been added
         * `MapOneToManyKeyValueRepo` class as implementation of `OneToManyKeyValueRepo` on top of `MutableMap` has been added
+
+
+## 0.1.0
+
+Inited :)